+++ collector-bitmovin-player/src/main/java/com/bitmovin/analytics/bitmovin/player/BitmovinPlayerCollector.java
@@ -7,7 +7,9 @@ import android.content.pm.PackageManager;
 import android.os.Build;
 import com.bitmovin.analytics.BitmovinAnalytics;
 import com.bitmovin.analytics.BitmovinAnalyticsConfig;
+import com.bitmovin.analytics.bitmovin.player.features.BitmovinFeatureFactory;
 import com.bitmovin.analytics.data.DeviceInformationProvider;
+import com.bitmovin.analytics.features.FeatureFactory;
 import com.bitmovin.player.BitmovinPlayer;
 
 public class BitmovinPlayerCollector extends BitmovinAnalytics {
@@ -30,12 +32,14 @@ public class BitmovinPlayerCollector extends BitmovinAnalytics {
     public void attachPlayer(BitmovinPlayer player) {
         DeviceInformationProvider deviceInformationProvider =
                 new DeviceInformationProvider(context, getUserAgent(context));
+        FeatureFactory featureFactory = new BitmovinFeatureFactory(this, player, context);
         BitmovinSdkAdapter adapter =
                 new BitmovinSdkAdapter(
                         player,
                         this.bitmovinAnalyticsConfig,
                         deviceInformationProvider,
-                        this.playerStateMachine);
+                        this.playerStateMachine,
+                        featureFactory);
 
         this.attach(adapter);