+++ collector/src/main/java/com/bitmovin/analytics/EventListener.kt
@@ -1,0 +1,3 @@
+package com.bitmovin.analytics
+
+interface EventListener