+++ collector/src/main/java/com/bitmovin/analytics/OnAnalyticsReleasingEventListener.kt
@@ -1,0 +1,10 @@
+package com.bitmovin.analytics
+
+interface OnAnalyticsReleasingEventSource {
+    fun addEventListener(listener: OnAnalyticsReleasingEventListener)
+    fun removeEventListener(listener: OnAnalyticsReleasingEventListener)
+}
+
+interface OnAnalyticsReleasingEventListener : EventListener {
+    fun onReleasing()
+}