+++ collector/src/main/java/com/bitmovin/analytics/features/FeatureConfig.kt
@@ -1,0 +1,5 @@
+package com.bitmovin.analytics.features
+
+abstract class FeatureConfig {
+    val enabled: Boolean? = null
+}