+++ collector/src/main/java/com/bitmovin/analytics/features/FeatureFactory.kt
@@ -1,0 +1,5 @@
+package com.bitmovin.analytics.features
+
+interface FeatureFactory {
+    fun createFeatures(): Collection<Feature<*>>
+}