+++ collector/src/main/java/com/bitmovin/analytics/features/segmenttracking/DownloadFinishedEvent.kt
@@ -1,0 +1,3 @@
+package com.bitmovin.analytics.features.segmenttracking
+
+data class DownloadFinishedEvent(val segment: Segment)