+++ collector/src/main/java/com/bitmovin/analytics/features/segmenttracking/SegmentTrackingConfig.kt
@@ -1,0 +1,7 @@
+package com.bitmovin.analytics.features.segmenttracking
+
+import com.bitmovin.analytics.features.FeatureConfig
+
+class SegmentTrackingConfig : FeatureConfig() {
+    val maxSegments: Int = 10
+}