+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Development
 
+- [Internal] `FeatureManager` allows dynamically adding and disabling features based on player support.
+- [Internal] `SegmentTrackingFeature` to always track the latest `n` downloaded elements
+- [Internal] `ErrorDetailsFeature` to track additional details in case an error occurs
+
 ## v1.22.0
 
 ### Added